616z_move_test_positive_cockpit_to_rex_component
